+++ Gruntfile.js
@@ -101,6 +101,7 @@ module.exports = function( grunt ) {
 					'assets/sass/woocommerce/extensions/bundles.css': 'assets/sass/woocommerce/extensions/bundles.scss',
 					'assets/sass/woocommerce/extensions/ship-multiple-addresses.css': 'assets/sass/woocommerce/extensions/ship-multiple-addresses.scss',
 					'assets/sass/woocommerce/extensions/advanced-product-labels.css': 'assets/sass/woocommerce/extensions/advanced-product-labels.scss',
+					'assets/sass/woocommerce/extensions/mix-and-match.css': 'assets/sass/woocommerce/extensions/mix-and-match.scss',
 					'assets/sass/woocommerce/woocommerce.css': 'assets/sass/woocommerce/woocommerce.scss',
 					'assets/sass/jetpack/jetpack.css': 'assets/sass/jetpack/jetpack.scss'
 				}]
@@ -291,6 +292,7 @@ module.exports = function( grunt ) {
 					'assets/sass/woocommerce/extensions/bundles.css',
 					'assets/sass/woocommerce/extensions/ship-multiple-addresses.css',
 					'assets/sass/woocommerce/extensions/advanced-product-labels.css',
+					'assets/sass/woocommerce/extensions/mix-and-match.css',
 					'assets/sass/woocommerce/woocommerce.css',
 					'assets/sass/admin/welcome-screen/welcome.css',
 					'assets/sass/jetpack/jetpack.css'