+++ inc/woocommerce/class-storefront-woocommerce.php
@@ -260,6 +260,14 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			}
 
 			/**
+			 * WooCommerce Mix and Match
+			 */
+			if ( $this->is_woocommerce_extension_activated( 'WC_Mix_and_Match' ) ) {
+				wp_enqueue_style( 'storefront-woocommerce-mix-and-match-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/mix-and-match.css', 'storefront-woocommerce-style' );
+				wp_style_add_data( 'storefront-woocommerce-mix-and-match-style', 'rtl', 'replace' );
+			}
+
+			/**
 			 * Checkout Add Ons
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_Checkout_Add_Ons' ) ) {